consolidate JoinAsDevice & JoinAsMember into single component
